refactor: decompose AccountDetails.tsx into sub-components
Extract 6 files from the 2,622-line AccountDetails component:
- account-details-types.ts: shared interfaces and types
- AutocompleteInput.tsx: autocomplete input component
- AccountHeader.tsx: header with rename, balance, search, actions
- TransactionForm.tsx: add transaction form (cash + investment)
- PendingOccurrences.tsx: pending scheduled transaction rows
- TransactionRow.tsx: single transaction row (view + edit modes)

Parent component retains all state and logic, passing data/callbacks
as props to presentational children. Default export preserved.
